requirements.txt
discord.py
requests
python-dotenv
pyjwt
schedule